remove gelu/broadcast shims. trace model in MPS+fp16. optimize trace. target macOS 13. try removing MPs via-fp32 cast for fp16 gelu. compute t_emb in fp16 to avoid a cast. add coremltools as a submodule. conversion finished (with drastic fp16 changes to coremltools), but loading fails since upsample seems to only support fp32.

+++ src/diffusers/models/embeddings.py
@@ -38,12 +38,12 @@ def get_timestep_embedding(
 
     half_dim = embedding_dim // 2
     exponent = -math.log(max_period) * torch.arange(
-        start=0, end=half_dim, dtype=torch.float32, device=timesteps.device
+        start=0, end=half_dim, dtype=torch.float16, device=timesteps.device
     )
     exponent = exponent / (half_dim - downscale_freq_shift)
 
     emb = torch.exp(exponent)
-    emb = timesteps[:, None].float() * emb[None, :]
+    emb = timesteps[:, None].to(dtype=emb.dtype) * emb[None, :]
 
     # scale embeddings
     emb = scale * emb
